Remove bold attribute from themes
The problem with setting bold is that terminals interpret it as a different colour. Not sure about GUIs. This creates inconsistencies, such as when an intended gre colour value appears as dt1 instead, because of the colour mapping. In the case of very high contrast themes, such as Caprice, the text is barely decipherable against the dark background.